+++ Camomile/camomile_do_not_use.ml
@@ -1,1 +1,0 @@
-[@@@ocaml.deprecated "this module is an empty. ignore it"]