+++ Camomile/toolslib/toolslib.ml
@@ -1,4 +1,0 @@
-open CamomileLibrary.Private
-module Unidata = Unidata.Make (Camomileconfig)
-module Charmap = Charmap.Configure (Camomileconfig)
-module Unimap = Unimap.Make (Camomileconfig)