+++ pkg/pkg.ml
@@ -1,3 +1,0 @@
-#use "topfind"
-
-#require "topkg-jbuilder.auto"