+++ src/config.ml
@@ -1,0 +1,2 @@
+include ConfigBase
+module Default = ConfigImpl