+++ src/toolslib/camomileconfig.ml
@@ -1,0 +1,6 @@
+(* configuration for tools *)
+
+let datadir = "../database"
+let localedir = "../locales"
+let charmapdir = "../charmaps"
+let unimapdir = "../mappings"