+++ src/toolslib/defaultModules.ml
@@ -1,0 +1,4 @@
+open Camomile.Private
+module Unidata = Unidata.Make (Camomileconfig)
+module Charmap = Charmap.Configure (Camomileconfig)
+module Unimap = Unimap.Make (Camomileconfig)